Move DashboardItem styles out of the top-level Dashboard styles.js file
