+++ src/components/Dashboard/DashboardItem/AnalysisItem.js
@@ -11,56 +11,53 @@ import { formatDate } from "@cyverse-de/ui-lib";
 
 import * as constants from "../constants";
 
-import ItemBase, { ItemAction } from "./ItemBase";
+import ItemBase, { ItemAction, getSectionClass } from "./ItemBase";
 
 class AnalysisItem extends ItemBase {
-    constructor({ section, content, height, width, classes }) {
+    constructor({ section, content, height, width }) {
         super({
             kind: constants.KIND_ANALYSES,
             content,
             section,
             height,
             width,
-            classes,
         });
     }
 
     static create(props) {
         const item = new AnalysisItem(props);
-        return item
-            .addActions([
-                <ItemAction
-                    ariaLabel="relaunch"
-                    key={`${constants.KIND_ANALYSES}-${props.content.id}-relaunch`}
-                >
-                    <Replay />
-                </ItemAction>,
-                <ItemAction
-                    ariaLabel="stop"
-                    key={`${constants.KIND_ANALYSES}-${props.content.id}-stop`}
-                >
-                    <Stop />
-                </ItemAction>,
-                <ItemAction
-                    ariaLabel="open details"
-                    key={`${constants.KIND_ANALYSES}-${props.content.id}-details`}
-                >
-                    <Info />
-                </ItemAction>,
-                <ItemAction
-                    ariaLabel="share"
-                    key={`${constants.KIND_ANALYSES}-${props.content.id}-share`}
-                >
-                    <Share />
-                </ItemAction>,
-                <ItemAction
-                    ariaLabel="go to output files"
-                    key={`${constants.KIND_ANALYSES}-${props.content.id}-outputs`}
-                >
-                    <FolderOpen />
-                </ItemAction>,
-            ])
-            .setSectionClass();
+        return item.addActions([
+            <ItemAction
+                ariaLabel="relaunch"
+                key={`${constants.KIND_ANALYSES}-${props.content.id}-relaunch`}
+            >
+                <Replay />
+            </ItemAction>,
+            <ItemAction
+                ariaLabel="stop"
+                key={`${constants.KIND_ANALYSES}-${props.content.id}-stop`}
+            >
+                <Stop />
+            </ItemAction>,
+            <ItemAction
+                ariaLabel="open details"
+                key={`${constants.KIND_ANALYSES}-${props.content.id}-details`}
+            >
+                <Info />
+            </ItemAction>,
+            <ItemAction
+                ariaLabel="share"
+                key={`${constants.KIND_ANALYSES}-${props.content.id}-share`}
+            >
+                <Share />
+            </ItemAction>,
+            <ItemAction
+                ariaLabel="go to output files"
+                key={`${constants.KIND_ANALYSES}-${props.content.id}-outputs`}
+            >
+                <FolderOpen />
+            </ItemAction>,
+        ]);
     }
 
     getOrigination(t) {
@@ -70,15 +67,10 @@ class AnalysisItem extends ItemBase {
         return [origination, formatDate(date.valueOf())];
     }
 
-    getAvatarIcon() {
-        if (!this.headerClass || this.avatarClass) {
-            this.setSectionClass();
-        }
+    getAvatarIcon(classes) {
+        const [, avatarClass] = getSectionClass(this.section, classes);
         return (
-            <BarChart
-                color="primary"
-                classes={{ colorPrimary: this.avatarClass }}
-            />
+            <BarChart color="primary" classes={{ colorPrimary: avatarClass }} />
         );
     }
